Enemy Die, EnemhDieDioeDieDie

+++ Assets/2. Scripts/Player/Hp/HealthSystem.cs
@@ -9,10 +9,10 @@ using UnityEngine;
         public event EventHandler OnHealthChanged;
         public event EventHandler OnDead;
 
-        private int healthMax;
-        private int health;
+        private float healthMax;
+        private float health;
 
-        public HealthSystem(int healthMax) {
+        public HealthSystem(float healthMax) {
             this.healthMax = healthMax;
             health = healthMax;
         }
@@ -21,7 +21,7 @@ using UnityEngine;
             return (float)health / healthMax;
         }
 
-        public void Damage(int amount) {
+        public void Damage(float amount) {
             health -= amount;
             if (health < 0) {
                 health = 0;
@@ -37,7 +37,7 @@ using UnityEngine;
             if (OnDead != null) OnDead(this, EventArgs.Empty);
         }
 
-        public void Heal(int amount) {
+        public void Heal(float amount) {
             health += amount;
             if (health > healthMax) {
                 health = healthMax;
